requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anykeystore==0.2
apex==0.9.10.dev0
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1733250440834/work
astunparse==1.6.3
attrs==25.1.0
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.tarfile==1.2.0
beautifulsoup4==4.12.3
bleach==6.2.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1733502965406/work
contourpy==1.3.0
cryptacular==1.6.2
cryptography==44.0.2
cycler==0.12.1
debugpy @ file:///home/conda/feedstock_root/build_artifacts/debugpy_1737269747879/work
decorator==5.1.1
defusedxml==0.7.1
docker-pycreds==0.4.0
docopt==0.6.2
docutils==0.21.2
einops==0.8.1
eval_type_backport==0.2.2
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1733569351617/work
fastjsonschema==2.21.1
filelock==3.17.0
flatbuffers==25.1.24
fonttools==4.56.0
fsspec==2025.2.0
ftfy==6.3.1
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.70.0
h5py==3.12.1
huggingface-hub==0.28.1
hupper==1.12.1
id==1.5.0
idna==3.10
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1737420181517/work
importlib_resources==6.5.2
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1719845459717/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1680185408135/work
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1733300866624/work
jeepney==0.9.0
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1733440914442/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1727163409502/work
jupyterlab_pygments==0.3.0
keras==3.8.0
keyring==25.6.0
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplot==0.1.9
matplotlib==3.9.4
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1733416936468/work
mdurl==0.1.2
mistune==3.1.1
ml-dtypes==0.4.1
more-itertools==10.6.0
munch==4.0.0
namex==0.0.8
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1733325553580/work
nh3==0.2.21
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.14.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1733203243479/work
pandas==2.2.3
pandocfilters==1.5.1
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1733271261340/work
PasteDeploy==3.1.0
pbkdf2==1.3
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1733301927746/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1733327343728/work
pillow==11.1.0
pipreqs==0.5.0
plaster==1.1.2
plaster-pastedeploy==1.0.1
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1733232627818/work
pretrainedmodels==0.7.4
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1737453357274/work
protobuf==5.29.3
psutil==6.1.1
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1733302279685/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl#sha256=92c32ff62b5fd8cf325bec5ab90d7be3d2a8ca8c8a3813ff487a8d2002630d1f
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1733569405015/work
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1736243443484/work
pyloco==0.0.139
pyparsing==3.2.1
pyramid==2.0.2
pyramid-mailer==0.15.1
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1733215673016/work
python-http-client==3.3.7
python3-openid==3.2.0
pytorch-gradcam==0.2.1
pytz==2025.1
PyYAML==6.0.2
pyzmq @ file:///home/conda/feedstock_root/build_artifacts/pyzmq_1738270954430/work
readme_renderer==44.0
referencing==0.36.2
regex==2024.11.6
repoze.sendmail==4.4.1
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.9.4
rpds-py==0.22.3
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.13.1
SecretStorage==3.3.3
sendgrid==6.11.0
sentry-sdk==2.20.0
setproctitle==1.3.4
SimpleWebSocketServer==0.1.2
six @ file:///home/conda/feedstock_root/build_artifacts/six_1733380938961/work
smmap==5.0.2
soupsieve==2.6
SQLAlchemy==2.0.37
stack_data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1733569443808/work
starkbank-ecdsa==2.2.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
threadpoolctl==3.6.0
timm==1.0.14
tinycss2==1.4.0
tokenizers==0.21.0
torch==1.13.1+cu116
torchaudio==0.13.1+cu116
torchvision==0.14.1+cu116
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1732615921868/work
tqdm==4.67.1
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1733367359838/work
transaction==5.0
transformers==4.48.3
translationstring==1.4
twine==6.1.0
typing==3.7.4.3
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1733188668063/work
tzdata==2025.1
urllib3==2.3.0
ushlex==0.99.1
velruse==1.1.1
venusian==3.1.1
wandb==0.19.5
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1733231326287/work
webencodings==0.5.1
WebOb==1.8.9
websocket-client==1.8.0
Werkzeug==3.1.3
wrapt==1.17.2
WTForms==3.2.1
wtforms-recaptcha==0.3.2
yarg==0.1.9
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1732827521216/work
zope.deprecation==5.1
zope.interface==7.2
zope.sqlalchemy==3.1
